+++ Commons/src/Commons.Core/Operation/OperationBase.cs
@@ -1,4 +1,5 @@
 ﻿using MudRunner.Commons.DataContracts.Operation;
+using System.Globalization;
 
 namespace MudRunner.Commons.Core.Operation
 {
@@ -11,6 +12,14 @@ namespace MudRunner.Commons.Core.Operation
         where TRequest : OperationRequestBase
         where TResponse : OperationResponseBase, new()
     {
+        /// <summary>
+        /// Class constructor.
+        /// </summary>
+        protected OperationBase()
+        {
+            CultureInfo.CurrentCulture = CultureInfo.InvariantCulture;
+        }
+
         /// <summary>
         /// Asynchronously, this method validates the request sent to operation.
         /// </summary>