+++ Commons/CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
  - Classes NewmarkMethodInput and NewmarkBetaMethodInput.
 ### Changed
  - Refactored numerical methods.
+ - Class OperationBase to set the current culture as invariant in the constructor.
 ### Fixed
  - Class OperationBase that was not processing the operation correctly.
  - Method 'AddErros' in class OperationResponseBase that was not adding errors when HttpStatusCode is not success.